(refactor) Mergerd 'login.js' to 'hotelSignin.js'
1. Merge 2files(login, hotelSignin) to hotelSignin.js
2. Delete login.js
3. Modify import in index.android.js
