+++ HotelAndroid/hotelSignin.js
@@ -2,55 +2,153 @@ import React, { Component } from 'react';
 import {
   View,
   StyleSheet,
-  TouchableWithoutFeedback,
   TouchableHighlight,
   Text,
-
+  TextInput,
+  AsyncStorage,
 } from 'react-native';
 import Register from './register';
 import Login from './login';
 
+const ACCESS_TOKEN = 'access_token';
 
 class HotelSignin extends Component {
   constructor(props) {
     super(props);
+
+    this.state = {
+      client_email : "",
+      password : "",
+      error : "",
+    }
   }
+
+// ----- the token related to client login ----- //
+
+  redirect(routeName, token) {
+      this.props.navigator.push({
+        id : routeName,
+        passProps : {
+          accessToken : token,//token
+        }
+      })
+    }
+
+  async storeToken(accessToken) {
+    try {
+      await AsyncStorage.setItem(ACCESS_TOKEN, accessToken);
+      this.getToken();
+    } catch(error) {
+      console.log("something went wrong")
+    }
+  }
+
+  async getToken(accessToken) {
+    try {
+      let token = await AsyncStorage.getItem(ACCESS_TOKEN);
+      console.log('token is : ' + token)
+    } catch(error) {
+      console.log("something went wrong")
+    }
+  }
+
+  async removeToken(accessToken) {
+    try {
+      await AsyncStorage.removeItem(ACCESS_TOKEN);
+      this.getToken();
+    } catch(error) {
+      console.log("something went wrong")
+    }
+  }
+
+  async onLoginPressed() {
+    try {
+      let response = await fetch('https://afternoon-beyond-22141.herokuapp.com/api/login', {
+        method: 'POST',
+        headers: {
+          'Accept': 'application/json',
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({
+          session:{
+            clientemail: this.state.email,
+            password: this.state.password,
+          }
+        })
+      });
+      let res = await response.text();
+      if (response.status >= 200 && response.status < 300) {
+        //Handle success
+        this.setState({error : ""});
+        let accessToken = res;
+        //On success we will store the access_token in the AsyncStorage
+        this.storeToken(accessToken);
+        console.log('res token : ' + accessToken);
+        this.redirect('check', accessToken);
+      } else {
+        //Handle error
+        let error = res;
+        throw error;
+      }
+    } catch(error) {
+      this.removeToken();
+      this.setState({error: error});
+      console.log("error " + error);
+      this.setState({showProgress: false});
+    }
+  }
+
+// -------------------------------------------- //
 
   _handlePress(routeName) {
     this.props.navigator.push({
       id : routeName,
     })
-    this.props.onChange(this.state.client_email);
+    // this.props.onChange(this.state.client_email);
   }
 
   render() {
     return (
-      <View>
-        <View style = {styles.container}>
-          <Login />
-        </View>
-        <View style={{flex : 2}}>
+      <View style = {styles.container}>
 
-          <TouchableHighlight
-            style={styles.register}
-            onPress = {this._handlePress.bind(this, 'register')}>
-            <Text style={styles.buttonText}>Register</Text>
-          </TouchableHighlight>
-        </View>
-        <View style={{flex : 3}}>
-          <TouchableHighlight
-            style={[styles.button, styles.facebook]}
-            onPress = {this._handlePress.bind(this, 'facebook')}>
-            <Text style={styles.buttonText}>Facebook</Text>
-          </TouchableHighlight>
-        </View>
-        <View style={{flex : 4}}>
-          <TouchableHighlight
-            style={[styles.button, styles.googleweb]}
-            onPress = {this._handlePress.bind(this, 'google')}>
-            <Text style={styles.buttonText}>google</Text>
-          </TouchableHighlight>
-        </View>
+        <Text style={styles.heading}>
+          💃 Enjoy Hotel-Reverse 💃
+        </Text>
+        <TextInput
+        onChangeText={ (text)=> this.setState({clientemail: text}) }
+        style={styles.input} placeholder="Email">
+        </TextInput>
+        <TextInput
+          onChangeText={ (text)=> this.setState({password: text}) }
+          style={styles.input}
+          placeholder="Password"
+          secureTextEntry={true}>
+        </TextInput>
+        <TouchableHighlight onPress={this.onLoginPressed.bind(this)} style={styles.button}>
+          <Text style={styles.buttonText}>
+            Login
+          </Text>
+        </TouchableHighlight>
+
+        <TouchableHighlight
+          style={styles.button}
+          onPress = {this._handlePress.bind(this, 'register')}>
+          <Text style={styles.buttonText}>Register</Text>
+        </TouchableHighlight>
+
+        <TouchableHighlight
+          style={[styles.button, styles.facebook]}
+          onPress = {this._handlePress.bind(this, 'facebook')}>
+          <Text style={styles.buttonText}>Facebook</Text>
+        </TouchableHighlight>
+        <TouchableHighlight
+          style={[styles.button, styles.googleweb]}
+          onPress = {this._handlePress.bind(this, 'google')}>
+          <Text style={styles.buttonText}>google</Text>
+        </TouchableHighlight>
+        <Text style={styles.error}>
+          {this.state.error}
+        </Text>
       </View>
     )
   }
@@ -64,7 +162,11 @@ let styles = StyleSheet.create({
   },
   container: {
     flex: 1,
-    backgroundColor : '#F5FCFF',
+    justifyContent: 'flex-start',
+    alignItems: 'center',
+    backgroundColor: '#F5FCFF',
+    padding: 10,
+    paddingTop: 80
   },
   content: {
     flex: 1,
@@ -79,17 +181,39 @@ let styles = StyleSheet.create({
     marginTop: 10,
     justifyContent: 'center'
   },
-  buttonText: {
-    color: '#fff',
+  input: {
+    height: 50,
+    marginTop: 10,
+    padding: 4,
     fontSize: 18,
-    alignSelf: 'center'
+    borderWidth: 1,
+    borderColor: '#48bbec'
   },
   button: {
-    height: 36,
-    flexDirection: 'row',
-    borderRadius: 8,
-    marginBottom: 10,
+    height: 50,
+    backgroundColor: '#48BBEC',
+    alignSelf: 'stretch',
+    marginTop: 10,
     justifyContent: 'center'
+  },
+  buttonText: {
+    fontSize: 22,
+    color: '#FFF',
+    alignSelf: 'center'
+  },
+  heading: {
+    fontSize: 30,
+  },
+  error: {
+    color: 'red',
+    paddingTop: 10
+  },
+  success: {
+    color: 'green',
+    paddingTop: 10
+  },
+  loader: {
+    marginTop: 20
   },
   pic: {
     width: 100,
@@ -122,11 +246,9 @@ let styles = StyleSheet.create({
   },
   'googleweb': {
     backgroundColor: '#ccc',
-    margin : 30
   },
   facebook: {
-    backgroundColor: '#3b5998',
-    margin : 30
+    backgroundColor: '#3b5998'
   },
   twitter: {
     backgroundColor: '#48BBEC'
